Accept NumericArray/PackedArray/ByteArray for all-numeric structs
When every field of a named struct shares the same numeric primitive
type (Point { x: i16, y: i16 }, etc.), the FromWolfram impl now also
accepts a single NumericArray, PackedArray, or ByteArray on the wire
and distributes the (widened) elements across the fields positionally.

Reuses numeric_in::read_fixed so the widening / element-count checks
and the no-truncation rule come from one place. The Association and
positional-Function paths are unchanged.

Also change wolfram_examples::Point's fields to i16 so the widening
rules are actually exercised end-to-end (Integer8 in widens, Real64
in rejected).

+++ wolfram-examples/src/lib.rs
@@ -37,8 +37,8 @@ pub fn duplicate(e: Expr) -> Expr {
 
 #[derive(Debug, Clone, PartialEq, ToWolfram, FromWolfram)]
 pub struct Point {
-    pub x: f64,
-    pub y: f64,
+    pub x: i16,
+    pub y: i16,
 }
 
 /// `Vec<f64>` fields are encoded as packed `NumericArray<Real64>` in WXF,
